requirements.txt
# Core
discord.py>=2.3,<3.0
langgraph>=0.2,<1.0
langchain-core>=0.3,<1.0
langchain-anthropic>=0.3,<1.0
anthropic>=0.40,<1.0
PyGithub>=2.1,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# Dev
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0
pytest-mock>=3.12,<4.0
ruff>=0.8,<1.0
